feat: use TypeScript + add tests + bump version
