+++ README.md
@@ -55,7 +55,7 @@ jobs:
     name: Capture Vercel preview URL
     runs-on: "ubuntu-latest"
     steps:
-      - uses: aaimio/vercel-preview-url-action@v1.1.1
+      - uses: aaimio/vercel-preview-url-action@v1.2.0
         id: vercel_preview_url
         with:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
