+++ package.json
@@ -1,7 +1,7 @@
 {
   "author": "Aaron Imming <aaim@protonmail.com>",
   "name": "vercel-preview-url-action",
-  "version": "1.1.1",
+  "version": "1.2.0",
   "description": "Capture a Vercel preview URL and feed it into your next GitHub action.",
   "main": "index.js",
   "husky": {
@@ -13,8 +13,8 @@
     "*.{js,md}": "prettier --write"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "npx @vercel/ncc build index.js --license licenses.txt --minify",
+    "build": "npx @vercel/ncc build index.ts --license licenses.txt --minify",
+    "test": "jest --verbose",
     "prepare": "husky install"
   },
   "repository": {
@@ -32,8 +32,12 @@
     "@octokit/action": "^4.0.10"
   },
   "devDependencies": {
+    "@types/jest": "^29.2.2",
     "husky": "^8.0.1",
+    "jest": "^29.2.2",
     "lint-staged": "^13.0.3",
-    "prettier": "^2.7.1"
+    "prettier": "^2.7.1",
+    "ts-jest": "^29.0.3",
+    "typescript": "^4.8.4"
   }
 }
